fix starwar mode
git-svn-id: svn://akdesign.dyndns.org/dtcyber/trunk@173 d8f3d347-7192-4e35-a67b-6d701703d080

+++ main.c
@@ -52,7 +52,7 @@
 **  Public Variables
 **  ----------------
 */
-i8 ppKeyIn;
+int ppKeyIn;
 bool emulationActive = TRUE;
 u32 cycles;
